refactor: split settings_window.rs into modular settings/ tabs [skip ci]
This fulfills task 6.13 by migrating the 1,666-line settings_window.rs
into a domain-specific settings/ directory. The commit fixes internal
constants and tests to ensure total compatibility and 100% logic preservation.
