+++ crates/katana-ui/src/settings/tabs/mod.rs
@@ -1,0 +1,6 @@
+pub mod behavior;
+pub mod font;
+pub mod layout;
+pub mod theme;
+pub mod updates;
+pub mod workspace;
